ml-services/trading/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
ta-lib==0.4.28
alpaca-py==0.11.3
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
asyncio==3.4.3
schedule==1.2.0
redis==5.0.1